+++ CHANGELOG.md
@@ -5,11 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.2] - 2025-07-01
+
+### Added
+
+- The `average_plot.py` script now also plots a linear fit with the average velocity.
+- Add new option `use_toroidal_hinge`, to allow simulating old non-toroidal hinge.
+
+### Changed
+
+- The `--visualize-follow` flag now has two choices `arms, kleisin`, tracking the SMC arms or kleisin respectively.
+- The `average_plot.py` script now converts to `nm` and `s` units when creating plots by default.
+- The `average_plot.py` script now removes `-1` values when finding the smallest common number of steps.
+
+### Fixed
+
+- The `process_displacement.py` script no longer raises an error when no `obstacle.lammpstrj` is present.
+
 ## [0.1.1] - 2025-06-17
 
 ### Added
 
-- LAMMPS script now print the SMC state (`APO`, `ATP`, `ADP`) to the screen.
+- LAMMPS script now prints the SMC state (`APO`, `ATP`, `ADP`) to the screen.
 
 ### Changed